Use the strongest available random source when shuffling EphIDs
We need to provide cryptographically secure random shuffling of EphIDs.

Looking at Python's implementation of random.shuffle()[1], we see that
(a) it is a straightforward implementation of Durstenfeld's shuffle
algorithm[2] (which is as unbiased as its random number source), and
(b) it allows for a custom `random` function to be passed in.

Hence, AFAICS, we should get a cryptographically secure shuffle as long
as we pass in a `random` function that itself is cryptographically
secure.

Quoting the documentation of the secrets module:

  The secrets module provides access to the most secure source of
  randomness that your operating system provides.

  class secrets.SystemRandom

    A class for generating random numbers using the highest-quality
    sources provided by the operating system.

Therefore, the best we can do is AFAICS to simply use the `random`
function available from secrets.SystemRandom class.

[1]: https://github.com/python/cpython/blob/v3.6.10rc1/Lib/random.py#L263
[2]: https://en.wikipedia.org/wiki/Fisher–Yates_shuffle#The_modern_algorithm
[3]: https://docs.python.org/3.6/library/secrets.html#random-numbers